Adds the is Cycle feature to stage
This change makes the stages to re-visit the same stage till
the breaking condition is met. As EM needs to record the metrics
for each cycle interval we need to run the same transistion multiple
times till the cycle needs to be stop based on a dynamic condition.

Usage of intermidiate stages is to keep track of a main stage and
intermidiate stages act as sub cycle stages, but here the requirement is
a bit different. Now we don't need multiple stages to be processed but
same stage to be scheduled multiple times so now we extend EM to support
cyclcing stages as well. So now we introduce the isCycle feature in EM
Stages.

PS: Lot of work needs to be done before we finalise the design of
cycling stages

Signed-off-by: bharathappali <abharath@redhat.com>